move ajax needed types into ajax code; add props types for container methods
